NMS-12261: fix datetime serialization to match JDK8
NOTE: The "JRE" mode will eventually go away and be replaced by
"COMPAT".  However, it has the advantage that it is supported by
JDKs 8 through 11 ("COMPAT" was introduced in JDK 9).

When we move to JDK 11+ entirely we should change the "JRE" ones to
"COMPAT".  (Or fix our parsing to handle both...)