feat: Simplify the container build from tarball
Remove build options from RPM and just install from a tarball. Just deploy to /opt/opennms and remove compatibilitiy
links to /usr/share/opennms to make it easier switching base images.
The java binary, doesn't set the capababilities to open network ports and getting access to the raw sockets. The
operational side needs to decide how the container is allowed to do that and shouldn't be baked in the container to
avoid security pitfalls.
It will make the contianer image also compatible to run in OpneShift environments.